# api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an event",
        "tags": [
          "events"
        ]
      },
      "get": {
        "operationId": "EventsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Events the user can read",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List events for the current user",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/join": {
      "post": {
        "operationId": "EventsController_join",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Join an event via invitation URL",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get an event by ID",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an event",
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event deleted (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/leave": {
      "post": {
        "operationId": "EventsController_leave",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Caller left the event (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Leave an event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/participants": {
      "get": {
        "operationId": "EventsController_getParticipants",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event participants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventParticipantResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List event participants",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/participants/{targetUserId}/access": {
      "put": {
        "operationId": "EventsController_updateParticipantAccess",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParticipantAccessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated member",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventParticipantResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a member access level",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/participants/{targetUserId}": {
      "delete": {
        "operationId": "EventsController_removeParticipant",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove a member from an event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/regenerate-url": {
      "post": {
        "operationId": "EventsController_regenerateInvitationUrl",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event with new invitation URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Regenerate the event invitation URL",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/cover/upload-url": {
      "post": {
        "operationId": "EventsController_createCoverUploadUrl",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload id with a presigned S3 PUT URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ImageUploadResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mint a presigned upload URL for the event cover image",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v2/events/{eventId}/cover": {
      "put": {
        "operationId": "EventsController_confirmCoverUpload",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImageUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event with the new cover",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EventResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Confirm an uploaded cover image and set it on the event",
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_removeCover",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cover removed (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove the event cover image",
        "tags": [
          "events"
        ]
      }
    }
  },
  "info": {
    "title": "Everglow API",
    "description": "Photo-sharing platform for events — HTTP API contract (OpenAPI 3)",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2026-06-03T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v2/auth/signin"
          }
        },
        "required": [
          "timestamp",
          "path"
        ]
      },
...
      "ImageUploadResponseDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid",
            "description": "Pass it back to confirm the upload"
          },
          "uploadUrl": {
            "type": "string",
            "maxLength": 2048,
            "description": "Presigned S3 PUT URL the client uploads bytes to"
          }
        },
        "required": [
          "uploadId",
          "uploadUrl"
        ]
      },
      "CreateImageUploadDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "example": "image/jpeg"
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 5242880,
            "example": 262144
          }
        },
        "required": [
          "contentType",
          "sizeBytes"
        ]
      },
      "ConfirmImageUploadDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid",
            "description": "The uploadId returned when the upload URL was minted"
          }
        },
        "required": [
          "uploadId"
        ]
      },
...
      "EventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "creatorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Who created the event; null once that account has been deleted. Not a permission: see accessLevel."
          },
          "invitationUrl": {
            "type": "string",
            "description": "Shareable invitation link composed from the stored invite token"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 2048,
            "description": "Short-lived presigned URL of the event cover image; null when none is set"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "date",
          "creatorId",
          "invitationUrl",
          "coverUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {}
      },
      "JoinEventDto": {
        "type": "object",
        "properties": {
          "invitationUrl": {
            "type": "string",
            "description": "Full invitation URL or invite token",
            "example": "https://events.everglow.app/invite/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
          }
        },
        "required": [
          "invitationUrl"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {}
      },
      "AccessLevel": {
        "type": "string",
        "enum": [
          "ORGANIZER",
          "PARTICIPANT",
          "VIEWER"
        ]
      },
      "EventParticipantResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "accessLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccessLevel"
              }
            ]
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 2048,
            "description": "Short-lived presigned URL of the member's avatar; null when none is set"
          }
        },
        "required": [
          "userId",
          "name",
          "accessLevel",
          "avatarUrl"
        ]
      },
      "UpdateParticipantAccessDto": {
        "type": "object",
        "properties": {
          "accessLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccessLevel"
              }
            ]
          }
        },
        "required": [
          "accessLevel"
        ]
      }
    },
    "responses": {
      "TooManyRequests": {
        "description": "Rate limit exceeded; retry after the number of seconds in the Retry-After header",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "meta"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Too many requests, please try again later"
                },
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable error code, when the error has one",
                  "example": "RATE_LIMIT_EXCEEDED"
                },
                "meta": {
                  "$ref": "#/components/schemas/ResponseMetaDto"
                }
              }
            }
          }
        }
      }
    }
  }
}